requirements.txt
google-genai>=0.6.0
pillow>=10.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.66.0